Move AG Proxy secret logic to activegate package; use context.Context
